backend/requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
cssbeautifier==1.15.1
defusedxml==0.7.1
Django==3.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
EditorConfig==0.12.4
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
inflection==0.5.1
jsbeautifier==1.15.1
json5==0.9.17
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.2.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tabulate==0.9.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
